added base interface [prototype]
+ interface
+ assets
+ manifest change
+ added loading screen to define screen widths with jquery because css
bugs out with  100% heights if your width is more than 100% so now we
scan the device, and then start fitting the elements to make sure it
fits. This should make it work on every device. Does not support big
screens atm ( media querys for background ) maybe ill get around to it
eventualy.